Extract true-on-policy MoE layer extensions and combine ordering
